agents/voice_agent/requirements.txt
openai-whisper
gtts
pydub
uvicorn
fastapi
python-multipart
